split modals in two
