+++ src/components/SaveModal.tsx
@@ -1,6 +1,6 @@
-// SaveModal.tsx
 import { useState } from "react";
 import { X, Save } from "lucide-react";
+import { Button } from "@/components/ui/button";
 
 interface SaveData {
   id: string;
@@ -59,8 +59,8 @@ export default function SaveModal({
   if (!isOpen) return null;
 
   return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md m-4">
+<div className="fixed inset-0 bg-black bg-opacity-50 flex overflow-y-auto p-4 z-50">
+      <div className="bg-white rounded-lg p-6 w-full max-w-md m-auto">
         <div className="flex justify-between items-center mb-4">
           <h2 className="text-xl font-bold">{title}</h2>
           <button
@@ -87,20 +87,22 @@ export default function SaveModal({
             />
           </div>
           <div className="flex gap-2 justify-end">
-            <button
+            <Button
               onClick={onClose}
-              className="px-4 py-2 bg-gray-300 text-gray-700 rounded-lg hover:bg-gray-400"
+              className="px-4 py-2"
+              variant="whiteSecondary"
             >
               Cancelar
-            </button>
-            <button
+            </Button>
+            <Button
               onClick={handleSave}
               disabled={!saveName.trim()}
-              className="px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600 disabled:opacity-50 disabled:cursor-not-allowed flex items-center gap-2"
+              className="px-4 py-2 flex items-center gap-2"
+              variant="secondary"
             >
               <Save className="w-4 h-4" />
               Salvar
-            </button>
+            </Button>
           </div>
         </div>
       </div>
